PDI-1450: Reorganize subcommand folders and packages, as well as update AddCommand calls as recommended
