+++ cmd/platform/export/export.go
@@ -5,7 +5,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+	http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package cmd
+package export
 
 import (
 	"fmt"
@@ -22,7 +22,7 @@ import (
 )
 
 // exportCmd represents the export command
-var exportCmd = &cobra.Command{
+var ExportCmd = &cobra.Command{
 	Use:   "export",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
@@ -37,8 +37,6 @@ to quickly create a Cobra application.`,
 }
 
 func init() {
-	platformCmd.AddCommand(exportCmd)
-
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
