+++ cmd/platform/platform.go
@@ -5,7 +5,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+	http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,16 +13,17 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package cmd
+package platform
 
 import (
 	"fmt"
 
+	"github.com/pingidentity/pingctl/cmd/platform/export"
 	"github.com/spf13/cobra"
 )
 
 // platformCmd represents the platform command
-var platformCmd = &cobra.Command{
+var PlatformCmd = &cobra.Command{
 	Use:   "platform",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
@@ -37,7 +38,7 @@ to quickly create a Cobra application.`,
 }
 
 func init() {
-	rootCmd.AddCommand(platformCmd)
+	PlatformCmd.AddCommand(export.ExportCmd)
 
 	// Here you will define your flags and configuration settings.
